Finished the basic restructure; added a working makefile

+++ Makefile
@@ -1,18 +1,27 @@
 # -*- Makefile -*-
 
-## Find all constituent RNW files:
-PARTS = $(filter-out index.Rmd, $(wildcard *.Rmd))
+## Find all constituent RMD files:
+CHAPTERS = $(filter-out index.Rmd, $(wildcard *.Rmd))
 
-#all: index
+VPATH = docs/
 
-index.Rmd: parts
-	Rscript -e "bookdown::render_book('$@')" 
-	
-parts: $(PARTS)
+all: index.html
+
+## Update book:
+index.html: index.Rmd chapters
+	Rscript -e "bookdown::render_book('$<')" 
+
+## Update chapters:
+chapters: $(CHAPTERS)
 	Rscript render.R $?
 	mv *.html docs/
-	touch parts
+	touch chapters
+
+## Update sections:
+%.Rmd: sections/%/* 
+	touch $@
 
 clean:
 	if [ -d ./_bookdown_files ]; then rm -r ./_bookdown_files; fi
-	rm -r docs/*
+	if [ -n "$(ls -A docs/*)" ]; then rm -r docs/*; fi
+	if [ -e willy_wallaby_from_wasatch.Rmd ]; then rm willy_wallaby_from_wasatch.Rmd; fi
